refactor: :recycle: Move GPU rasterizer into separate struct
