+++ src/lib.rs
@@ -5,7 +5,6 @@ pub mod read;
 pub mod render;
 pub mod scene;
 pub mod surface;
-pub mod trivial_rasterizer;
 pub mod tui;
 
 /// Module for learning ASCII rendering
